Moved the TagInput file inside add contact directly and converted to pug
